services/whisper/requirements.txt
# 1. Whisper 核心與轉錄
faster-whisper>=1.0.0
opencc-python-reimplemented
librosa>=0.10.0
soundfile>=0.12.1

# 2. Web 服務
fastapi>=0.100.0
uvicorn>=0.25.0
pydantic>=2.0.0
python-dotenv
